Move WebMapService createFromXML, isVersion1_1_1, isVersion1_3_0 to WebMapServiceFactoryHelper, WebMapServiceFactory class can not be loaded in unit tests
